[New] added new test cases for transitions
